requirements.txt
azure-functions
onnxruntime
opencv-python-headless
numpy
Pillow